requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.3
django-qr-code==4.0.1
djangorestframework==3.15.1
pydantic==2.6.4
pydantic_core==2.16.3
segno==1.6.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
